+++ designs/unice/tests/shared.test.mjs
@@ -1,10 +1,10 @@
 // This file is auto-generated | Any changes you make will be overwritten.
-import { Unice } from './dist/index.mjs'
+import { Unice } from '../src/index.mjs'
 
 // Shared tests
 import { testPatternConfig } from '../../../tests/designs/config.mjs'
 import { testPatternDrafting } from '../../../tests/designs/drafting.mjs'
-import { testPatternSampling } from '../../../tests/designs/sampling.mjs'
+//import { testPatternSampling } from '../../../tests/designs/sampling.mjs'
 
 // Test config
 testPatternConfig(Unice)
@@ -13,4 +13,4 @@ testPatternConfig(Unice)
 testPatternDrafting(Unice, false)
 
 // Test sampling - Change the second parameter to `true` to log errors
-testPatternSampling(Unice, false)
+//testPatternSampling(Unice, false)