+++ CHANGELOG.md
@@ -649,6 +649,12 @@
 
  - Migrated from Rollup to Esbuild for all builds
 
+### unice
+
+#### Added
+
+ - Unice is an underwear pattern
+
 ### ursula
 
 #### Changed